refactor: implement CreatureMovedEvent and update event handling for creature movement
